requirements.txt
flask
reportlab
python-docx
PyPDF2
groq
python-dotenv
pymupdf
sentence-transformers
faiss-cpu
numpy
